requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
filelock==3.9.0
Flask==3.0.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.4.0
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
idna==3.4
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lightning-utilities==0.10.1
Mako==1.3.1
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
mlflow==2.9.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
packaging==23.2
pandas==2.2.0
Pillow==9.3.0
protobuf==4.25.2
pyarrow==14.0.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.3
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.28.1
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.25
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchmetrics==1.3.0.post0
torchsummary==1.5.1
torchvision==0.16.2+cu118
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.13
waitress==2.1.2
websocket-client==1.7.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
